+++ tests/testthat/teardown.R
@@ -10,5 +10,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and limitations under the License.
 
+unlink(getOption("bcmaps.data_dir"), recursive = TRUE, force = TRUE)
 options(bc_data_dir)
 options(silence_update_message_value)